pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "solanaeasy"
version = "0.1.0"
description = "A Stripe para Solana — pagamentos blockchain sem saber que é blockchain"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
authors = [
    { name = "Equipe SolanaEasy — UNICAMP", email = "solanaeasy@dev.com" },
]
keywords = ["solana", "blockchain", "payments", "sdk", "usdc"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "click>=8.0",
]

# Dependências Solana separadas — opcionais pois o SDK pode rodar
# somente chamando o backend (sem acesso direto à rede Solana)
[project.optional-dependencies]
solana = [
    "solders>=0.21",
    "solana>=0.35",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-mock>=3.12",
    "respx>=0.21",          # Mock para httpx
    "ruff>=0.4",
    "mypy>=1.10",
    "python-dotenv>=1.0",
]
all = ["solanaeasy[solana,dev]"]

[project.scripts]
solanaeasy = "solanaeasy.cli:cli"

[project.urls]
Homepage = "https://github.com/solanaeasy/solanaeasy-python"
Documentation = "https://docs.solanaeasy.dev"

[tool.hatch.build.targets.wheel]
packages = ["solanaeasy"]

# ── Ruff (linter + formatter) ─────────────────────────────────────
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM"]
ignore = ["E501"]

# ── Mypy (type checker) ───────────────────────────────────────────
[tool.mypy]
python_version = "3.11"
strict = true
ignore_missing_imports = true

# ── Pytest ────────────────────────────────────────────────────────
[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
